moving propmt interactor
