+++ src/main/java/use_case/chat/PrompInteractor.java
@@ -1,4 +1,4 @@
-package data_access.gpt_api_calls;
+package use_case.chat;
 
 import entity.bot.MasterYodaFactory;
 import entity.bot.NormalAIFactory;
